requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
discord.py==1.6.0
idna==2.10
multidict==5.1.0
oauthlib==3.1.0
PySocks==1.7.1
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
tornado==6.1
tweepy==3.10.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
yarl==1.6.3
